+++ package.json
@@ -1,5 +1,5 @@
 {
-  "homepage": "https://diasky.github.io/portfolio/",
+  "homepage": "https://caamillo.github.io/",
   "name": "camillo-portfolio",
   "version": "0.1.0",
   "private": true,
